Recognize gitlessness within mounted FSes
When the current directory belongs to a filesystem other than /, git outputs a different error message after failing its upward search for a repository. This commit handles the different message, treating it similarly to the original message. This also fixes a failing spec on systems where the tmp directory is mounted as tmpfs, such as Arch.